backend/pyproject.toml
[project]
name = "blitz-backend"
version = "0.1.0"
description = "Blitz — AI-powered marketing pipeline backend"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "litellm>=1.40.0",
    "chromadb>=0.5.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "aiosqlite>=0.22.1",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
